+++ src/pages/blog/[blogSlug].tsx
@@ -7,6 +7,8 @@ import {
   PortableTextComponentProps,
 } from "@portabletext/react";
 import { useNextSanityImage } from "next-sanity-image";
+import { useTheme } from "styled-components";
+import { uniqBy } from "lodash/fp";
 
 import Layout from "../../components/Layout";
 import CMSClient, {
@@ -25,7 +27,7 @@ import Grid, { GridArea } from "../../components/Grid";
 import MaxWidth from "../../components/MaxWidth/MaxWidth";
 import Box from "../../components/Box";
 import { Heading, P, Span } from "../../components/Typography";
-// import CopyLinkButton from "../../components/Button/CopyLinkButton";
+import CopyLinkButton from "../../components/Button/CopyLinkButton";
 import { getCTAHref } from "../../utils/portableText/resolveInternalHref";
 import { OmitKeepDiscriminated } from "../../utils/generics";
 import ButtonAsLink from "../../components/Button/ButtonAsLink";
@@ -33,6 +35,9 @@ import { BasePortableTextProvider } from "../../components/PortableText";
 import { BlogJsonLd } from "../../browser-lib/seo/getJsonLd";
 import CMSVideo from "../../components/CMSVideo";
 import { getSeoProps } from "../../browser-lib/seo/getSeoProps";
+import MobileBlogFilters from "../../components/MobileBlogFilters";
+import OakLink from "../../components/OakLink";
+import BlogCategoryList from "../../components/BlogCategoryList";
 import Circle from "../../components/Circle";
 import { extractNumberEnvVar } from "../../utils/configHelper";
 
@@ -42,6 +47,7 @@ export type SerializedBlog = Omit<BlogPost, "date"> & {
 
 export type BlogPageProps = {
   blog: SerializedBlog;
+  categories: { title: string; slug: string }[];
 };
 
 // When we get the JSON portable text it doesn't have the same field names as
@@ -221,7 +227,7 @@ const logMissingPortableTextComponents: MissingComponentHandler = (
 };
 
 const BlogDetailPage: NextPage<BlogPageProps> = (props) => {
-  const { blog } = props;
+  const { blog, categories } = props;
 
   const formattedDate = new Date(blog.date).toLocaleDateString("en-GB", {
     day: "numeric",
@@ -242,6 +248,8 @@ const BlogDetailPage: NextPage<BlogPageProps> = (props) => {
         imageUrlBuilder.width(1400).height(700).fit("crop").crop("center"),
     }
   );
+  const theme = useTheme();
+  const HEADER_HEIGHT = theme.header.height;
 
   return (
     <Layout
@@ -253,20 +261,33 @@ const BlogDetailPage: NextPage<BlogPageProps> = (props) => {
       })}
       $background="white"
     >
+      <MobileBlogFilters categoryListProps={{ categories }} withBackButton />
       <MaxWidth>
         <Grid $ph={[12, 0]}>
-          <GridArea $colSpan={[12, 7]}>
+          <GridArea $order={[0, 2]} $colSpan={[12, 3]}>
+            <Box
+              $display={["none", "block"]}
+              $position={[null, "sticky"]}
+              $top={[null, HEADER_HEIGHT]}
+              $pt={[48, 72]}
+            >
+              <Heading tag="h3" $font="body-3">
+                Categories
+              </Heading>
+              <BlogCategoryList $mt={24} categories={categories} />
+            </Box>
+          </GridArea>
+          <GridArea $order={[0, 1]} $colSpan={[12, 2]} />
+          <GridArea $order={[1, 0]} $colSpan={[12, 7]}>
             <Flex
-              $mt={[40, 80]}
+              $mt={[40, 72]}
               $justifyContent="space-between"
               $flexDirection={["column", "row"]}
             >
-              <Heading
-                tag={"h2"}
-                $font={["heading-7"]}
-                $color="black" // change to "hyperlink" when it becomes a link
-              >
-                {blog.category.title}
+              <Heading tag={"h2"} $color="hyperlink" $font={["heading-7"]}>
+                <OakLink page="blog-index" category={blog.category.slug}>
+                  {blog.category.title}
+                </OakLink>
               </Heading>
               <Span $font={"body-3"} $mt={[8, 0]}>
                 {formattedDate}
@@ -275,30 +296,36 @@ const BlogDetailPage: NextPage<BlogPageProps> = (props) => {
             <Heading $mt={12} $font={["heading-5", "heading-4"]} tag={"h1"}>
               {blog.title}
             </Heading>
-            <Flex $alignItems={"center"} $mt={16}>
-              {blog.author.image && (
-                <Circle $mr={12} $overflow={"hidden"} size={56}>
-                  <CMSImage
-                    image={{
-                      altText: blog.author.image.altText,
-                      isPresentational: true,
-                      asset: blog.author.image.asset,
-                    }}
-                  />
-                </Circle>
-              )}
-              <Box>
-                <Heading tag="h2" $font={"heading-7"} $mr={40}>
-                  {blog.author.name}
-                </Heading>
-                {blog.author.role && (
-                  <P $mt={4} $font={"body-3"} $color={"oakGrey4"}>
-                    {blog.author.role}
-                  </P>
+            <Flex
+              $alignItems={"center"}
+              $mt={16}
+              $mr={[20, 0]}
+              $justifyContent={["space-between", "left"]}
+            >
+              <Flex $alignItems={"center"}>
+                {blog.author.image && (
+                  <Circle $mr={12} $overflow={"hidden"} size={56}>
+                    <CMSImage
+                      image={{
+                        altText: blog.author.image.altText,
+                        isPresentational: blog.author.image.isPresentational,
+                        asset: blog.author.image.asset,
+                      }}
+                    />
+                  </Circle>
                 )}
-              </Box>
-              {/* TODO: add more UI for copy link button */}
-              {/* <CopyLinkButton /> */}
+                <Box $mr={[0, 40]}>
+                  <Heading tag="h2" $font={"heading-7"}>
+                    {blog.author.name}
+                  </Heading>
+                  {blog.author.role && (
+                    <P $mt={4} $font={"body-3"} $color={"oakGrey4"}>
+                      {blog.author.role}
+                    </P>
+                  )}
+                </Box>
+              </Flex>
+              <CopyLinkButton />
             </Flex>
             <Box $mt={[48]}>
               <BasePortableTextProvider>
@@ -342,6 +369,12 @@ export const getStaticProps: GetStaticProps<BlogPageProps, URLParams> = async (
     previewMode: isPreviewMode,
   });
 
+  const blogResults = await CMSClient.blogPosts();
+  const categories = uniqBy(
+    "title",
+    blogResults.map((blogResult) => blogResult.category)
+  ).sort((a, b) => (a.title < b.title ? -1 : 1));
+
   if (!blogResult) {
     return {
       notFound: true,
@@ -355,6 +388,7 @@ export const getStaticProps: GetStaticProps<BlogPageProps, URLParams> = async (
 
   return {
     props: {
+      categories,
       blog,
       isPreviewMode,
     },